fix: improvement in file workflow
