+++ examples/ui/frontend/src/components/ui/script-viewer.tsx
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import { ChevronRight, Code } from "lucide-react";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { vscDarkPlus } from "react-syntax-highlighter/dist/esm/styles/prism";
+import { getLanguage } from "@/lib/utils";
 
 interface ScriptViewerProps {
   code?: string;
@@ -33,39 +34,6 @@ const ScriptViewer: React.FC<ScriptViewerProps> = ({
       return code.length > maxLength ? `${code.substring(0, maxLength)}...` : code;
     };
     return truncateCode(code);
-  };
-
-  const getLanguage = (language?: string): string => {
-    if (!language) return "javascript";
-    
-    const languageMap: Record<string, string> = {
-      "python": "python",
-      "javascript": "javascript",
-      "js": "javascript",
-      "typescript": "typescript",
-      "ts": "typescript",
-      "bash": "bash",
-      "shell": "bash",
-      "sh": "bash",
-      "php": "php",
-      "java": "java",
-      "cpp": "cpp",
-      "c++": "cpp",
-      "c": "c",
-      "go": "go",
-      "rust": "rust",
-      "ruby": "ruby",
-      "sql": "sql",
-      "html": "html",
-      "css": "css",
-      "json": "json",
-      "yaml": "yaml",
-      "yml": "yaml",
-      "markdown": "markdown",
-      "md": "markdown"
-    };
-    
-    return languageMap[language.toLowerCase()] || "javascript";
   };
 
   const renderExpandedContent = () => {
